Initial Commit - Chef for Non-Tech
This is the initial slide deck for the Chef for Non-Tech class that I
taught for Chef Staff.

This is very much a work in progress.

It contains the following sections:

1)  Understanding DevOps

2) Typical Infrastructure

3)  ChefSpeak

4)  Hosted Chef

5)  Chef Server

6)  Creating an Apache Workbook/writing a web server recipe

    Resources/Recipes/Cookbooks

7)  Test Kitchen Demo

8)  Knife Berks

9)  Community Cookbooks

10)  Chef Delivery Workflow

____________________________

I am currently preparing slides for two different use cases:

Single Node
Multiple Nodes

Also, students have requested an explanation of the differences in
various products:

   Ansible
   Puppet
   Chef
   Docker
   VMs
   Azure
   AWS

And a clearer explanation of BLDR - still working on that.